# openapi/management-data-needs.yaml
openapi: 3.0.0
info:
  title: Data Needs Management API
  version: 1.0.0
  description: REST API for managing data needs stored in the application database, providing CRUD operations.

servers:
  - url: 'http://localhost:9090/management'
    description: Main API server

paths:
  /data-needs:
    post:
      summary: Create a new DataNeed
      tags:
        - DataNeeds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataNeed'
      responses:
        '201':
          description: Data need created
        '409':
          description: Data need already exists or invalid input

    get:
      summary: Get all DataNeeds
      tags:
        - DataNeeds
      responses:
        '200':
          description: An array of data needs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataNeed'

  /data-needs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a DataNeed by ID
      tags:
        - DataNeeds
      responses:
        '200':
          description: Data need found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataNeed'
        '404':
          description: Data need not found

    put:
      summary: Update a DataNeed by ID
      tags:
        - DataNeeds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataNeed'
      responses:
        '200':
          description: Data need updated
        '400':
          description: Mismatched IDs or invalid data
        '404':
          description: Data need not found

    delete:
      summary: Delete a DataNeed by ID
      tags:
        - DataNeeds
      responses:
        '204':
          description: Data need deleted
        '404':
          description: Data need not found

components:
  schemas:
    DataNeed:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/DataType'
        granularity:
          $ref: '#/components/schemas/Granularity'
        durationStart:
          type: integer
        durationOpenEnd:
          type: boolean
        durationEnd:
          type: integer
        transmissionInterval:
          type: integer
        sharedDataIds:
          type: array
          items:
            type: string
        serviceName:
          type: string
      required:
        - id
        - description
        - type
        - granularity
        - durationStart
        - durationOpenEnd

    DataType:
      type: string
      enum:
        - HISTORICAL_VALIDATED_CONSUMPTION_DATA
        - FUTURE_VALIDATED_CONSUMPTION_DATA
        - AIIDA_NEAR_REALTIME_DATA
        - ACCOUNTING_POINT_MASTER_DATA

    Granularity:
      type: string
      enum:
        - PT5M
        - PT10M
        - PT15M
        - PT30M
        - PT1H
        - P1D
        - P1M
        - P1Y